+++ config/services.yaml
@@ -6,7 +6,7 @@ parameters:
     app_locales: en|nl|nl_NL|nl_BE|es|fr|de|pl|it|pt_BR
     app.notifications.email_sender: anonymous@example.com
     doctrine.orm.entity_manager.class: Bolt\Doctrine\TranslatableEntityManager
-    stof_doctrine_extensions.listener.translatable.class: Bolt\EventListener\PreTranslatableListener
+    stof_doctrine_extensions.listener.translatable.class: Bolt\Event\Listener\PreTranslatableListener
     bolt.table_prefix: bolt_
     bolt.backend_url: /bolt
 
@@ -27,7 +27,7 @@ services:
     # this creates a service per class whose id is the fully-qualified class name
     Bolt\:
         resource: '../src/*'
-        exclude: '../src/{Entity,Migrations,Tests,Kernel.php}'
+        exclude: '../src/{Entity,Exception,Kernel.php}'
 
     # controllers are imported separately to make sure services can be injected
     # as action arguments even if you don't extend any base controller class
@@ -35,13 +35,12 @@ services:
         resource: '../src/Controller'
         tags: ['controller.service_arguments']
 
-    Bolt\EventListener\ContentFillListener:
+    Bolt\Event\Listener\ContentFillListener:
         tags:
             - { name: doctrine.event_listener, event: postLoad }
 
-
-    Twig\Extension\StringLoaderExtension: ~ # @todo this one is autoloaded?
-    Doctrine\ORM\Query\Expr: ~ # @todo this one is autoloaded?
+    Twig\Extension\StringLoaderExtension: ~
+    Doctrine\ORM\Query\Expr: ~
 
     # Needed for SetContent from bolt/bolt
     Bolt\Storage\Query\Query:
