+++ src/Controller/Backend/MediaController.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Bolt\Controller\Backend;
 
 use Bolt\Configuration\FileLocations;
-use Bolt\Content\MediaFactory;
 use Bolt\Controller\TwigAwareController;
+use Bolt\Factory\MediaFactory;
 use Doctrine\Common\Persistence\ObjectManager;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Component\Finder\Finder;
