+++ src/Entity/Field.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Bolt\Entity;
 
 use ApiPlatform\Core\Annotation\ApiResource;
-use Bolt\Content\FieldType;
+use Bolt\Configuration\Content\FieldType;
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Mapping\Annotation as Gedmo;
 use Symfony\Component\Serializer\Annotation\Groups;
