+++ src/Factory/MediaFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Bolt\Content;
+namespace Bolt\Factory;
 
 use Bolt\Configuration\Config;
 use Bolt\Configuration\FileLocations;
